After much work
User buttons can be added and layouts can be added as well.

There is a bug wheb adding a new layout. It will save it to the db, but initially fails.

the ubPosInLayout currently gets it's value from button's Id. However this should change so that buttons are contiguos within a layout.